[UPDATE] Enemy move, speed 기능
적 움직임과 스피드, 생성 확률 기능 수정

+++ Assets/Scripts/Enemy/EnemyData.cs
@@ -6,6 +6,6 @@ public class EnemyData : ScriptableObject
     [Header("Enemy")]
     public string enemyName = "Enemy";
     public int enemyMaxHp = 400;
-    public int enemySpeed = 1;
+    public float enemySpeed = 1f;
     public Sprite sprite;
 }
